Parse markdown only once.
By the time `marked` invokes the `renderer.paragraph` callback, it has already processed the content and translated anything it could parse to HTML. After splitting on block markers, we should therefore avoid re-parsing the content as (sanitized) Markdown), because that will escape the HTML markup produced by the initial parse.

To see the problem, try a brew like this:

```
{{test
![Goblin](http://i.imgur.com/ET6Hp6D.png)
}}
```

Without this change, the `<img...>` tag produced by the first parse is escaped into `&lt;img...`, and therefore the image is not displayed. After the patch, the image renders as expected.