Outline app into a separate file
Main motivation is to separate app definition and "app.listen"
pieces of code so port 8000 is not busy every time you launch
tests.

Also added dependency on `app-module-path` in order to unify
require paths: when launched under Jest and under regular
`npm server.js`, `NODE_PATH` is different which means that the
same path to an included file can not be found in on or the
other setup by default. This change adds root directory of a
projec to `NODE_PATH` and for Jest we can do the same through
`modulePaths` or `moduleDirectories` so paths within a file
are the same for both "entry points".