Fix #773 Wide Class Table snippet missing header
- There should be a "—Spell Slots per Spell Level—" header slightly above
the 1st-9th spell slot headers.

- This fix attempts to add this in as simple a manner as possible; it
involves a tiny bit of CSS, 1 new line to the snippet, and 1 moved line
within the snippet. The new snippet uses the heading row for the new
header and re-styles the first "normal" row to look like a header, and
subsequent rows to have the correct alternating background colours.
- The standalone stylesheet has been updated to reflect this new CSS as
well.

- The new snippet includes the `twoHeadingRows` class to be fully
backwards-compatible with documents using the old snippet.