Create .frame class for Class Tables
Add a `.frame` class for Class Tables so that the frame and background can be toggled on or off in a similar fashion to the new monster stat block.

This will allow a class table to be used as a "third caster archetype" table such as referenced in PR #1079.  Adds consistency with Monster blocks, too.