Create Solution for simplified shaped image wrap.
This adds a pattern for creating macro substitutions in
mustacheInjectInline to create complex style properties that depend on
information in the moustache-impacted token.

Creates $swl and $swr macros for images to use shape-outside.

$swl is changed to shape-outside: url(image_url), float: left
$swr is changed to shape-outside: url(image_url), float: right

in both cases image_url is the href value from lastToken.

The FAQ is also updated with how to use this macro or perform the tasks
manually.