License Snippet Dropdown update
This extends on the work done in PR 3090 to solve #1976

I reorganized things a little bit to use sub menus and files per
licensing organization. It makes it a bit easier to work around in
though it's still awkward with the Creative Commons
To Do:
  Creative Commons "deeds"
  Creative Commons image snippets

Maybe To Do:
  Various Publisher community licenses, such as the AGE Creators
Alliance, Super-Powered by M&M, various Drive Through RPG Content
licenses, etc.