Change text wrap make to default to shape-margin
This switches the default wrapping context for the text wrap macros from
image-threshold to margin which is a more common pattern. Image
threshold has been retained in the a new pair of macros, twtl and twtr.

If a margin value is provided, it assumes the unit of measure is
expected to be provided.