Create setter and getter methods for cursor position
Setter takes a 'view', grabs the current cursor position, and dumps it into localStorage with a related key.

Getter takes a 'view', loads the related localStorage key, and *if* an editor instance is focused, sets the cursor position in it.  The CM instance must have focus on it already-- related code is in next commit.