Change active input ref system
Any tag can be turned into the active input and the inputref is dynamically assigned to that input.  So the "new tag" input isn't really that different from a "edit existing tag" input.

Some whitespace and formatting changes as well.  Stepped away from stashed changes for 8 days so this isn't the best commit msg.