Working "Magic Button"
This feature works in two parts.

This first part is by adding a prioxy-passthrough for resources. This enables the code to get around XSS issues we were previously having,.

The second part clones the first page, massages URLs in it's <Links>, <style>, and <img> tags to use the
L to load the resources then generate two thumbnails. One is 115px tall and the other is set at 50% the page height. Currently, only the smaller of the two is stored as a new BrewStub field.