Exporting updates.
This changes the PDF print option to an Export menu with PDF as the first choice and CBZ as the second.

CBZ Export is functional, but not quite perfect. There are a few border effects that do not seem to work correctly with the image export.

As the export progresses the left margin will vanish while it processes to account for the the canvas convertor interpretting the margin. As a plus this functions as a progress meter. :)