TOC uses headings for links instead of page ID
this will allow a more precise toc that will jump to the titles instead of the page, it works on web and printed pdf and will avoid to have to regenerate the TOC if the page number changed.

Obviously there regenerating the TOC is needed if the titles are modified, but that would be necessary anyway since you would have a different text in TOC and titles.