Improve keyboard navigation
Add options for:

- arrow key and tab navigation,
- 'Enter' or 'Space' to edit a tag
- 'Enter', 'Tab', or ',' to submit a new tag
- 'Escape' to exit a tag without submitting
- 'Delete' to delete a tag

Tabbing from outside the component sends focus to the new tag input, from which you can shift-tab or arrowLeft to move focus through the tags.

Tabbing or shift-tabbing through the last tag or past the new tag input sends focus out of the component to wherever the next focus point is.