Rework combobox.jsx to functional component
This is barebones features of the combobox, but written as a functional component.  It features a text input and a div containing a list of buttons with the different options included.  It functionally updates the brew metadata props and loads them when the component initializes.

Many features of the original combobox are still missing.

Light re-styling.

Small HTML changes to metadataEditor.  Change the list of options to buttons rather than divs with onClick attributes.

Change how options are created and passed into Combobox component.