Add moment.js import & use to set date strings
Remove a bunch of lines of code determining "hours ago" elapsed time ranges, replacing it with moment.js functions.  Moment is already installed in the repo.

Change the `title` strings to include formatted date/time strings and a little more info.

Add a "restore from..." header to the dropdown to provide a little context clue for what the buttons do.