refactor/restructure of snippetbar.jsx
Most changes here are to the nodes that contain the snippet buttons and menus, and the snippets themselves.  The snippetbar now contains two subgroups:  snippets, and the "other ones".  The "other ones" are the smaller buttons, called 'tools' and includes the editors (which is a somewhat awkward grouping).

The 'tools' are further grouped into history tools, code tools (aka CM tools), and the editor buttons.

Snippets get revamped into html `popover`s with CSS anchor positioning.  They become click-only activated.