Automatically re-check for scripts
Adding a separate `keepText` field for the `updateBrew()` API might be a bandaid for something that should be looked at more deeply as a separate refactor, considering `updateBrew()` is configured to just return the stub and not the whole document.

For now, re-scanning for script tags after updating can be as simple as just re-looking up the brew.