Change "googleDrive" button to new Dialog
Change the Google Drive transfer toggle so that it opens a modal, rather than be a straight up transfer button.  The modal UI gives the opportunity to add more services in the future-- multiple services can be listed and chosen from.  Additionally, the modal gives more information about what is actually happening (more can be added, or linked to in the FAQ).

Further, any errors incurred can be displayed directly in the same modal before closing the modal.

There is stuff in this commit that might not make sense right away, particularly around states.