add line widgets and classes to error'd lines
Largely chatGPT'd.  I've tested this a bit, and it works the vast majority of the time.  however, there was one test I can't reproduce right now that left one line highlighted with the error line class (but not the line widget) when it shouldn't have.  It was copy/pasting a bunch of tags around different lines that caused it.