Add error panel above code editor
Previous commits added line widgets to highlight errors right where they are, but this commit adds a panel above the code editor that appears when htmlErrors exist.  The panel shows a list of all errors, including the line number, and a button to scroll to each error.

It also fixes the issue where the markdown validation would run on every keypress (an issue introduced in earlier commits).  Now it again only does the validation on save, and then on every press if errors are already exist (up until they don't).