add ability to swap panes
adds a button on the divider bar that allows editor and preview panes to be swapped.  preserves ability to brew and source jump between panes.  **but liveScroll linking doesn't work**, which is what I'm working on next.  The issue is that the Panes themselves are not swapping, only the contents within those panes.  So the Pane props are not aligned with their children props/methods.