Rework SplitPane and Pane
Removes scroll functions from SplitPane and narrows the component to only handling layout.  Adds the ability to swap the panes.

Buttons on the divider are now passed in as props, along with their functionality.  SplitPane can now be used more generically, such as in the Vault Page.

Touches a lot of files that use SplitPane to update props etc.