update the Vault page search pane
Brings the Vault page styling more inline with the rest of the UI:  darker background, matching fonts and sizes.   It also switches the validation of the renderer checkbox from a brittle CSS method to just using component state.

Scrollbar styling is moved out to it's own global style file as  a mixin.

Form styling is moved out to a more generic style file (and will ultimately be moved to the global styles folder and improved-- and metadata editor will likely adopt it as well).